Calculate indices of Group items by row and column
Previously, the indices were calculated as (row index * number of
columns) + column index. Now, the new indices are calculated
consecutively, first by row and then by column. This change allows the
maximum number of items in a group to reach the square of the maximum
number of items in a layer.